backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
pydantic>=2.0.0
openai>=1.30.0
python-dotenv>=1.0.0
tavily-python>=0.3.0
httpx>=0.27.0
